{"schema":"decodex/commit/1","summary":"Split graph fact storage helpers into write, query, and supersession modules after strict validation.","authority":"manual"}
